Eliminado id de todas las clases del dominio, se ha cambiado por una superclase de Spring Data AbstractPersistable. Los id ahora son de tipo Long. Se ha cambiado la clase test de certification, ahora termina y da correcto. Dojo y Bootsrap ahora se importan a través de Maven y Webjars.

+++ src/main/java/com/acme/repository/AuxiliarRepository.java
@@ -30,8 +30,8 @@ public interface AuxiliarRepository {
 	public void removeLanguage(Language l);
 	public void removeOffice(Office o);
 	public void removeMethodPay(MethodPay mp);
-	public Office getOfficeById(Integer id);
-	public Language getLanguageById(Integer id);
-	public Country getCountryById(Integer id);
-	public MethodPay getMethodPayById(Integer id);
+	public Office getOfficeById(Long id);
+	public Language getLanguageById(Long id);
+	public Country getCountryById(Long id);
+	public MethodPay getMethodPayById(Long id);
 }
