+++ src/main/java/com/acme/services/CertificationService.java
@@ -9,10 +9,10 @@ import com.acme.model.certification.FamilyProfessional;
 
 public interface CertificationService{
 	public List<Certification> getAllCertification();
-	public Certification getCertificationById(Integer id);
-	public void removeCertificationById(Integer id);
+	public Certification getCertificationById(Long id);
+	public void removeCertificationById(Long id);
 	public void updateCertification(Certification cert);
 	public void createCertification(Certification cert);
 	public List<FamilyProfessional> getAllFamilyProfessional();
-	public FamilyProfessional getFamilyProfessionalById(Integer id);
+	public FamilyProfessional getFamilyProfessionalById(Long id);
 }
